Refactor param structs to centralize in models package
